% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{crandeps}
\alias{crandeps}
\alias{crandeps-package}
\title{CRAN Package Dependencies}
\description{
Query CRAN package dependencies, calculate recursive dependencies, or
the whole dependency DAG.
}

