% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ crandeps }
\alias{ crandeps }
\alias{ crandeps }
\title{ CRAN Package Dependencies }
\description{
Query CRAN (and other CRAN-like) package dependencies, nice and easy,
  calculate recursive dependencies, or the whole dependency DAG.
}
